--- src/test/helpers.js
@@ -1,7 +1,0 @@
-const delete_sample_post = test_ref => {
-  test_ref.remove()
-}
-
-module.exports = {
-  delete_sample_post
-}
